+++ scripts/contributions/models.ts
@@ -66,6 +66,7 @@ export type MenuLocations =
 	| 'explorer/context'
 	| 'extension/context'
 	| 'git.commit'
+	| 'git.changes'
 	| 'menuBar/edit/copy'
 	| 'scm/title'
 	| 'scm/sourceControl'