+++ src/webviews/apps/plus/timeline/timeline.ts
@@ -202,7 +202,7 @@ export class GlTimelineApp extends GlAppHost<State> {
 			<gl-repo-button-group
 				aria-label="Visualize Repository History"
 				.connectIcon=${false}
-				.hasMultipleRepositories=${false /* TODO support repositories switching */}
+				.hasMultipleRepositories=${this.state.repositories.openCount > 1}
 				.icon=${false}
 				.repository=${repo}
 				.source=${{ source: 'timeline' } as const}